user-service/requirements.txt
django>=6.0
djangorestframework
djangorestframework-simplejwt
mysql-connector-python
